+++ CHANGES.md
@@ -40,6 +40,7 @@ Changes in 12.0
 - Add FileDelete() function in FileUpload.fnc.php
 - Security: use FileDelete() instead of unlink(), program wide
 - Students Search: remove unused code for `$extra['array_function']` in Search.inc.php
+- Add AddonUnzip() & AddOnZipCanUnzip() functions in Addons.fnc.php
 
 Changes in 11.8.2
 -----------------